+++ packages/astro/test/client/sdk.test.ts
@@ -48,14 +48,6 @@ describe('Sentry client SDK', () => {
       );
     });
 
-    it('sets the runtime tag on the isolation scope', () => {
-      expect(getIsolationScope().getScopeData().tags).toEqual({});
-
-      init({ dsn: 'https://public@dsn.ingest.sentry.io/1337' });
-
-      expect(getIsolationScope().getScopeData().tags).toEqual({ runtime: 'browser' });
-    });
-
     describe('automatically adds integrations', () => {
       it.each([
         ['tracesSampleRate', { tracesSampleRate: 0 }],