+++ packages/remix/src/index.server.ts
@@ -1,11 +1,6 @@
 import { applySdkMetadata } from '@sentry/core';
 import type { NodeClient, NodeOptions } from '@sentry/node';
-import {
-  getDefaultIntegrations as getDefaultNodeIntegrations,
-  init as nodeInit,
-  isInitialized,
-  setTag,
-} from '@sentry/node';
+import { getDefaultIntegrations as getDefaultNodeIntegrations, init as nodeInit, isInitialized } from '@sentry/node';
 import type { Integration } from '@sentry/types';
 import { logger } from '@sentry/utils';
 
@@ -193,8 +188,6 @@ export function init(options: RemixOptions): NodeClient | undefined {
   const client = nodeInit(options as NodeOptions);
 
   instrumentServer(options);
-
-  setTag('runtime', 'node');
 
   return client;
 }