+++ packages/remix/test/index.client.test.ts
@@ -43,12 +43,4 @@ describe('Client init()', () => {
   it('returns client from init', () => {
     expect(init({})).not.toBeUndefined();
   });
-
-  it('sets runtime on scope', () => {
-    expect(SentryReact.getIsolationScope().getScopeData().tags).toEqual({});
-
-    init({ dsn: 'https://public@dsn.ingest.sentry.io/1337' });
-
-    expect(SentryReact.getIsolationScope().getScopeData().tags).toEqual({ runtime: 'browser' });
-  });
 });